+++ src/app-config/dicts.ts
@@ -139,6 +139,40 @@ export function getALlDicts () {
         max: 5,
       },
     },
+    cnki: {
+      lang: '11000000',
+      selectionLang: {
+        english: true,
+        chinese: true,
+        japanese: false,
+        korean: false,
+        french: false,
+        spanish: false,
+        deutsch: false,
+        others: false,
+      },
+      defaultUnfold: {
+        english: true,
+        chinese: true,
+        japanese: true,
+        korean: true,
+        french: true,
+        spanish: true,
+        deutsch: true,
+        others: true,
+      },
+      preferredHeight: 300,
+      selectionWC: {
+        min: 1,
+        max: 100,
+      },
+      options: {
+        dict: true,
+        senbi: true,
+        seneng: true,
+        // digests: true,
+      }
+    },
     cobuild: {
       lang: '10000000',
       selectionLang: {