+++ src/content/components/WordEditor/index.tsx
@@ -107,7 +107,7 @@ export class WordEditor extends React.PureComponent<WordEditorProps & { t: Trans
     this.getRelatedWords()
     const word = this.props.editorWord
     if (word.context && !word.trans) {
-      translateCtx(word.context, this.props.ctxTrans)
+      translateCtx(word.context || word.title, this.props.ctxTrans)
         .then(trans => {
           if (trans) {
             const word = this.props.editorWord
@@ -158,6 +158,7 @@ export class WordEditor extends React.PureComponent<WordEditorProps & { t: Trans
               <a
                 href='https://github.com/crimx/ext-saladict/issues/117'
                 target='_blank'
+                rel='nofollow noopener noreferrer'
               > Why?</a>
             </label>
             <textarea rows={5}